move performConfiguredRequest to strict files and update imports
