+++ src/extensionPages/extensionPagePlatform.ts
@@ -24,13 +24,13 @@ import type { UUID } from "@/types/stringTypes";
 import {
   traces,
   clearExtensionDebugLogs,
+  performConfiguredRequestInBackground,
 } from "@/background/messenger/strict/api";
 import { PlatformBase } from "@/platform/platformBase";
 import type { Nullishable } from "@/utils/nullishUtils";
 import type { SanitizedIntegrationConfig } from "@/integrations/integrationTypes";
 import type { NetworkRequestConfig } from "@/types/networkTypes";
 import type { RemoteResponse } from "@/types/contract";
-import { performConfiguredRequestInBackground } from "@/background/messenger/api";
 import integrationRegistry from "@/integrations/registry";
 import { performConfiguredRequest } from "@/background/requests";
